Adding tests for indexed primitives.
