+++ phantom-jdk8/src/test/scala/com/outworkers/phantom/jdk8/indexed/IndexedRoundtripTests.scala
@@ -1,0 +1,67 @@
+/*
+ * Copyright 2013 - 2017 Outworkers Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.outworkers.phantom.jdk8.indexed
+
+import java.time.{LocalDate, LocalDateTime, OffsetDateTime, ZoneOffset, ZonedDateTime}
+
+import com.outworkers.phantom.jdk8.tables._
+import com.outworkers.phantom.builder.primitives.Primitive
+import com.outworkers.util.samplers.Sample
+import org.scalacheck.Gen
+import org.scalatest.{Assertion, FlatSpec, Matchers}
+import org.scalatest.prop.GeneratorDrivenPropertyChecks
+
+class IndexedRoundtripTests extends FlatSpec with Matchers with GeneratorDrivenPropertyChecks {
+
+  private[this] val genLower: Int = -100000
+  private[this] val genHigher: Int = -genLower
+
+  implicit override val generatorDrivenConfig: PropertyCheckConfiguration = {
+    PropertyCheckConfiguration(minSuccessful = 300)
+  }
+
+
+  def roundtrip[T](fn: T => T)(
+    implicit ev: Primitive[T],
+    sm: Sample[T]
+  ): Assertion = {
+    forAll(Sample.generator[T], protocolGen) { (sample, version) =>
+      ev.deserialize(ev.serialize(sample, version), version) shouldEqual fn(sample)
+    }
+  }
+
+  def roundtrip[T](gen: Gen[T])(fn: T => T)(implicit ev: Primitive[T]): Assertion = {
+    forAll(gen, protocolGen) { (sample, version) =>
+      ev.deserialize(ev.serialize(sample, version), version) shouldEqual fn(sample)
+    }
+  }
+
+  it should "serialize and deserialize ZonedDateTime instances using indexed primitives" in {
+    roundtrip[ZonedDateTime](_.withZoneSameInstant(ZoneOffset.UTC))
+  }
+
+  it should "serialize and deserialize OffsetDateTime instances using indexed primitives" in {
+    roundtrip[OffsetDateTime](_.withOffsetSameInstant(ZoneOffset.UTC))
+  }
+
+  it should "serialize and deserialize LocalDate instances using indexed primitives" in {
+    roundtrip[LocalDate](identity)
+  }
+
+  it should "serialize and deserialize LocalDateTime instances using indexed primitives" in {
+    roundtrip[LocalDateTime](identity)
+  }
+}
